ai-node-svr/code/src/tools/ds_tsapp_0003_DsLabelToolSvr/requirements.txt
oss2
PyYAML
requests==2.23.0
scipy
scikit-image
python-dotenv
whatimage
matplotlib
opencv-python==4.11.0.86
pillow
minio
flasgger
colorlog
apscheduler